wc: moved some '*-autocomplete' component to deprecated folder
