+++ src/webcomponents/commons/filters/deprecated/ethnicity-autocomplete.js
@@ -15,11 +15,10 @@
  */
 
 import {LitElement, html} from "lit";
-import UtilsNew from "../../../core/utilsNew.js";
-import "../../commons/forms/select-token-filter.js";
+import "../../forms/select-token-filter.js";
 
-
-export default class FileNameAutocomplete extends LitElement {
+// Rodiel 06-05-2022 - DEPRECATED: use catalog-distinct-autocomplete now.
+export default class EthnicityAutocomplete extends LitElement {
 
     createRenderRoot() {
         return this;
@@ -55,26 +54,18 @@ export default class FileNameAutocomplete extends LitElement {
 
     getDefaultConfig() {
         return {
-            placeholder: "eg. samples.tsv, phenotypes.vcf...",
             limit: 10,
-            fields: item => ({
-                name: item.name,
-                Format: item.format ?? "N/A",
-                Size: UtilsNew.getDiskUsage(item.size)
-            }),
             source: (params, success, failure) => {
                 const page = params?.data?.page || 1;
-                const name = params?.data?.term ? {id: "~/" + params?.data?.term + "/i"} : null;
+                const ethnicity = params?.data?.term ? {ethnicity: "~/" + params?.data?.term + "/i"} : null;
                 const filters = {
                     study: this.opencgaSession.study.fqn,
                     limit: this._config.limit,
                     count: false,
                     skip: (page - 1) * this._config.limit,
-                    type: "FILE",
-                    include: "id,name,format,size",
-                    ...name
+                    ...ethnicity
                 };
-                this.opencgaSession.opencgaClient.files().search(filters)
+                this.opencgaSession.opencgaClient.individuals().distinct("ethnicity", filters)
                     .then(response => success(response))
                     .catch(error => failure(error));
             },
@@ -94,4 +85,4 @@ export default class FileNameAutocomplete extends LitElement {
 
 }
 
-customElements.define("file-name-autocomplete", FileNameAutocomplete);
+customElements.define("ethnicity-autocomplete", EthnicityAutocomplete);
