+++ src/webcomponents/commons/filters/deprecated/proband-id-autocomplete.js
@@ -15,10 +15,10 @@
  */
 
 import {LitElement, html} from "lit";
-import "../../commons/forms/select-token-filter.js";
+import "../../forms/select-token-filter.js";
 
-
-export default class EthnicityAutocomplete extends LitElement {
+// Rodiel 06-05-2022 - DEPRECATED: use catalog-distinct-autocomplete now.
+export default class ProbandIdAutocomplete extends LitElement {
 
     createRenderRoot() {
         return this;
@@ -55,17 +55,20 @@ export default class EthnicityAutocomplete extends LitElement {
     getDefaultConfig() {
         return {
             limit: 10,
+            /* fields: item => ({
+                name: item.id
+            }),*/
             source: (params, success, failure) => {
                 const page = params?.data?.page || 1;
-                const ethnicity = params?.data?.term ? {ethnicity: "~/" + params?.data?.term + "/i"} : null;
+                const proband = params?.data?.term ? {proband: "~/" + params?.data?.term + "/i"} : null;
                 const filters = {
                     study: this.opencgaSession.study.fqn,
                     limit: this._config.limit,
                     count: false,
                     skip: (page - 1) * this._config.limit,
-                    ...ethnicity
+                    ...proband
                 };
-                this.opencgaSession.opencgaClient.individuals().distinct("ethnicity", filters)
+                this.opencgaSession.opencgaClient.clinical().distinct("proband.id", filters)
                     .then(response => success(response))
                     .catch(error => failure(error));
             },
@@ -80,9 +83,10 @@ export default class EthnicityAutocomplete extends LitElement {
                 .value="${this.value}"
                 @filterChange="${e => this.onFilterChange("id", e.detail.value)}">
             </select-token-filter>
+
         `;
     }
 
 }
 
-customElements.define("ethnicity-autocomplete", EthnicityAutocomplete);
+customElements.define("proband-id-autocomplete", ProbandIdAutocomplete);
